Reorganise files
- create providers/ directory and move relevant components there
- move App.tsx to src/
- remove CatchallRedirect file, move it's
component definition to Router.tsx
