--- src/components/CatchallRedirect.tsx
@@ -1,6 +1,0 @@
-import { Redirect, useParams } from 'react-router'
-
-export default function CatchallRedirect() {
-  const route = useParams<{ route: string }>()['route']
-  return <Redirect to={'/p/' + route} />
-}
